Added support for Route URI Models
You can now add a model to your URI and that model is available inside
the route controller as well as in the view.

/my/{model:App\Models\User}/uri

The part in the URI when the model is defined will be passed in as a
parameter to the models construct.

View tags are {model:method_name}

- Also did some optimisation with routes passing data in and out of
controllers (now much more efficient).